Add template and include for "legacy" policies.
In order to observe the old "number headings" and "sub-number points"
format of the older era policies, we need some extra CSS.

To ensure we don't kill the new-style policies by mistake I've
duplicated them and made forward changes to isolate them. This should
have a short lifetime and we can purge it when no longer required.