Alter the seeding policy to reward semi-finalists from previous event
While the World Ranking Policy is does a good job of dealing with regional
changes to playing strength over time (particularly between World Cups) it can
result in a relegation of teams who performed well but have simply not played
since.

We don't want to revert to the former system where recent form is entirely
ignored for 4 year old data, however we do want to assure nations that making
it into the semi-finals is an excellent achievement and provide some longer
term reward for that.

This amendment will bypass the ranking points for the top-4 finishers at an
event and the fill in behind them based on ranking points.