Proposed changes to the FIT Player Eligibility Policy.
These changes are being proposed to clarify a number of contentious
issues in the existing policy. The two major issues being:

U18 exemption
-------------
This revision removes the blanket exclusion that prevents the policy
from being applied to players under the age of 18.

The original intention of the exemption is to protect minors from being
bound to a decision made prior to adulthood and possibly outside of
their control.

It was never intended to permit youth players to represent a Member
without firstly meeting the eligibility criteria.

This revision protects the original intent (a player may freely choose
who they will Represent once they have turned 18) however they must at
all times be eligible for the Member they are Representing.

Nationality across borders
--------------------------
Some nationalities span multiple countries. For example British
Citizenship covers England, Scotland, Wales, Jersey, Guernsey and some
of Ireland.

This revision clarifies how the policy is to apply to a player who is a
Legal National or Naturalised Citizen in this situation.