Reword for "Event Delivery team"
The goal here is to make this a high-performance team with specialists
in a number of roles. Having a single position description is inflexible
as we need a complementary set of skills and experiences.

Focus is to make this an assembly of hand picked talent, possibly
head-hunted based on what we're looking for or need to bolster.